Consistency around access to VALID_FILE_REGEX
