+++ src/current.ts
@@ -12,12 +12,11 @@ import {
   runStringMigration,
   serializeHeader,
 } from "./migration";
+import { idFromFilename, isMigrationFilename } from "./migrationFilename";
 import { withClient, withTransaction } from "./pg";
 import { ParsedSettings } from "./settings";
 import pgMinify = require("pg-minify");
 import { DbCurrent } from "./interfaces";
-
-export const VALID_FILE_REGEX = /^([0-9]+)(-[-_a-zA-Z0-9]*)?\.sql$/;
 
 async function statOrNull(path: string): Promise<Stats | null> {
   try {
@@ -89,24 +88,6 @@ export async function getCurrentMigrationLocation(
     exists,
     // stats,
   };
-}
-
-function idFromFilename(file: string): number {
-  const matches = VALID_FILE_REGEX.exec(file);
-  if (!matches) {
-    throw new Error(
-      `Invalid current migration filename: '${file}'. File must follow the naming 001.sql or 001-message.sql, where 001 is a unique number (with optional zero padding) and message is an optional alphanumeric string.`,
-    );
-  }
-  const [, rawId, _message] = matches;
-  const id = parseInt(rawId, 10);
-
-  if (!id || !isFinite(id) || id < 1) {
-    throw new Error(
-      `Invalid current migration filename: '${file}'. File must start with a (positive) number, could not coerce '${rawId}' to int.`,
-    );
-  }
-  return id;
 }
 
 export async function readCurrentMigration(
@@ -321,7 +302,7 @@ export async function writeCurrentMigration(
     const files = await fsp.readdir(location.path);
     for (const file of files) {
       if (
-        VALID_FILE_REGEX.test(file) &&
+        isMigrationFilename(file) &&
         !file.startsWith(".") &&
         file.endsWith(".sql") &&
         !filenamesWritten.includes(file)
